refactor: move large header lines to cpp
