+++ src/cellular_automata.h
@@ -17,27 +17,9 @@ namespace CellularAutomata
 * @param p
 * @returns The number of neighbors that are alive, a value between 0 and 8.
 */
-inline int neighbor_count( const std::vector<std::vector<int>> &cells, const point &size,
-                           const point &p )
-{
-    int neighbors = 0;
-    for( int ni = -1; ni <= 1; ni++ ) {
-        for( int nj = -1; nj <= 1; nj++ ) {
-            const point n( p + point( ni, nj ) );
-
-            // These neighbors are outside the bounds, so they can't contribute.
-            if( n.x < 0 || n.x >= size.x || n.y < 0 || n.y >= size.y ) {
-                continue;
-            }
-
-            neighbors += cells[n.x][n.y];
-        }
-    }
-    // Because we included ourself in the loop above, subtract ourselves back out.
-    neighbors -= cells[p.x][p.y];
-
-    return neighbors;
-}
+int neighbor_count( const std::vector<std::vector<int>> &cells,
+                    const point &size,
+                    const point &p );
 
 /**
 * Generate a cellular automaton using the provided parameters.
@@ -54,53 +36,12 @@ inline int neighbor_count( const std::vector<std::vector<int>> &cells, const poi
 * @param stasis_limit Alive cells with > statis_limit neighbors stay alive
 * @returns The width x height grid of cells. Each cell is a 0 if dead or a 1 if alive.
 */
-inline std::vector<std::vector<int>> generate_cellular_automaton( const point &size,
-                                  const int alive, const int iterations, const int birth_limit, const int stasis_limit )
-{
-    std::vector<std::vector<int>> current( size.x, std::vector<int>( size.y, 0 ) );
-    std::vector<std::vector<int>> next( size.x, std::vector<int>( size.y, 0 ) );
-
-    // Initialize our initial set of cells.
-    for( int i = 0; i < size.x; i++ ) {
-        for( int j = 0; j < size.y; j++ ) {
-            current[i][j] = x_in_y( alive, 100 );
-        }
-    }
-
-    for( int iteration = 0; iteration < iterations; iteration++ ) {
-        for( int i = 0; i < size.x; i++ ) {
-            for( int j = 0; j < size.y; j++ ) {
-                // Skip the edges--no need to complicate this with more complex neighbor
-                // calculations, just keep them constant.
-                if( i == 0 || i == size.x - 1 || j == 0 || j == size.y - 1 ) {
-                    next[i][j] = 0;
-                    continue;
-                }
-
-                // Count our neighors.
-                const int neighbors = neighbor_count( current, size, point( i, j ) );
-
-                // Dead and > birth_limit neighbors, so become alive.
-                if( ( current[i][j] == 0 ) && ( neighbors > birth_limit ) ) {
-                    next[i][j] = 1;
-                }
-                // Alive and > statis_limit neighbors, so stay alive.
-                else if( ( current[i][j] == 1 ) && ( neighbors > stasis_limit ) ) {
-                    next[i][j] = 1;
-                }
-                // Else, die.
-                else {
-                    next[i][j] = 0;
-                }
-            }
-        }
-
-        // Swap our current and next vectors and repeat.
-        std::swap( current, next );
-    }
-
-    return current;
-}
+std::vector<std::vector<int>> generate_cellular_automaton(
+                               const point &size,
+                               const int alive,
+                               const int iterations,
+                               const int birth_limit,
+                               const int stasis_limit );
 } // namespace CellularAutomata
 
 #endif // CATA_SRC_CELLULAR_AUTOMATA_H
