+++ src/generic_factory.h
@@ -848,43 +848,14 @@ inline void optional( const JsonObject &jo, const bool was_loaded, const std::st
  * Reads a string and stores the first byte of it in `sym`. Throws if the input contains more
  * or less than one byte.
  */
-inline bool one_char_symbol_reader( const JsonObject &jo, const std::string &member_name, int &sym,
-                                    bool )
-{
-    std::string sym_as_string;
-    if( !jo.read( member_name, sym_as_string ) ) {
-        return false;
-    }
-    if( sym_as_string.size() != 1 ) {
-        jo.throw_error( member_name + " must be exactly one ASCII character", member_name );
-    }
-    sym = sym_as_string.front();
-    return true;
-}
+bool one_char_symbol_reader( const JsonObject &jo, const std::string &member_name, int &sym,
+                             bool );
 
 /**
  * Reads a UTF-8 string (or int as legacy fallback) and stores Unicode codepoint of it in `symbol`.
  * Throws if the inputs width is more than one console cell wide.
  */
-inline bool unicode_codepoint_from_symbol_reader( const JsonObject &jo,
-        const std::string &member_name, uint32_t &member, bool )
-{
-    int sym_as_int;
-    std::string sym_as_string;
-    if( !jo.read( member_name, sym_as_string, false ) ) {
-        // Legacy fallback to integer `sym`.
-        if( !jo.read( member_name, sym_as_int ) ) {
-            return false;
-        } else {
-            sym_as_string = string_from_int( sym_as_int );
-        }
-    }
-    uint32_t sym_as_codepoint = UTF8_getch( sym_as_string );
-    if( mk_wcwidth( sym_as_codepoint ) != 1 ) {
-        jo.throw_error( member_name + " must be exactly one console cell wide", member_name );
-    }
-    member = sym_as_codepoint;
-    return true;
-}
+bool unicode_codepoint_from_symbol_reader( const JsonObject &jo,
+        const std::string &member_name, uint32_t &member, bool );
 
 #endif // CATA_SRC_GENERIC_FACTORY_H
