+++ src/itype.h
@@ -890,12 +890,8 @@ struct itype {
         itype();
         virtual ~itype();
 
-        int damage_min() const {
-            return count_by_charges() ? 0 : damage_min_;
-        }
-        int damage_max() const {
-            return count_by_charges() ? 0 : damage_max_;
-        }
+        int damage_min() const;
+        int damage_max() const;
 
         // used for generic_factory for copy-from
         bool was_loaded = false;
@@ -1036,69 +1032,25 @@ struct itype {
 
         FlagsSetType item_tags;
 
-        std::string get_item_type_string() const {
-            if( tool ) {
-                return "TOOL";
-            } else if( comestible ) {
-                return "FOOD";
-            } else if( container ) {
-                return "CONTAINER";
-            } else if( armor ) {
-                return "ARMOR";
-            } else if( book ) {
-                return "BOOK";
-            } else if( gun ) {
-                return "GUN";
-            } else if( bionic ) {
-                return "BIONIC";
-            } else if( ammo ) {
-                return "AMMO";
-            }
-            return "misc";
-        }
+        std::string get_item_type_string() const;
 
         // Returns the name of the item type in the correct language and with respect to its grammatical number,
         // based on quantity (example: item type “anvil”, nname(4) would return “anvils” (as in “4 anvils”).
         std::string nname( unsigned int quantity ) const;
 
         // Allow direct access to the type id for the few cases that need it.
-        const itype_id &get_id() const {
-            return id;
-        }
+        const itype_id &get_id() const;
 
-        bool count_by_charges() const {
-            return stackable_ || ammo || comestible;
-        }
+        bool count_by_charges() const;
 
-        int charges_default() const {
-            if( tool ) {
-                return tool->def_charges;
-            } else if( comestible ) {
-                return comestible->def_charges;
-            } else if( ammo ) {
-                return ammo->def_charges;
-            }
-            return count_by_charges() ? 1 : 0;
-        }
+        int charges_default() const;
 
-        int charges_to_use() const {
-            if( tool ) {
-                return static_cast<int>( tool->charges_per_use );
-            }
-            return 1;
-        }
+        int charges_to_use() const;
 
         // for tools that sub another tool, but use a different ratio of charges
-        int charge_factor() const {
-            return tool ? tool->charge_factor : 1;
-        }
+        int charge_factor() const;
 
-        int maximum_charges() const {
-            if( tool ) {
-                return tool->max_charges;
-            }
-            return 0;
-        }
+        int maximum_charges() const;
         bool can_have_charges() const;
 
         /**
