+++ src/point.cpp
@@ -6,6 +6,24 @@
 
 #include "cata_utility.h"
 
+
+point point::rotate( int turns, const point &dim ) const
+{
+    assert( turns >= 0 );
+    assert( turns <= 4 );
+
+    switch( turns ) {
+        case 1:
+            return { dim.y - y - 1, x };
+        case 2:
+            return { dim.x - x - 1, dim.y - y - 1 };
+        case 3:
+            return { y, dim.x - x - 1 };
+    }
+
+    return *this;
+}
+
 std::string point::to_string() const
 {
     std::ostringstream os;
