+++ src/sdl_utils.h
@@ -17,43 +17,9 @@ using color_pixel_function_map = std::unordered_map<std::string, color_pixel_fun
 
 color_pixel_function_pointer get_color_pixel_function( const std::string &name );
 
-inline SDL_Color adjust_color_brightness( const SDL_Color &color, int percent )
-{
-    if( percent <= 0 ) {
-        return { 0x00, 0x00, 0x00, color.a };
-    }
+SDL_Color adjust_color_brightness( const SDL_Color &color, int percent );
 
-    if( percent == 100 ) {
-        return color;
-    }
-
-    return SDL_Color{
-        static_cast<Uint8>( std::min( color.r *percent / 100, 0xFF ) ),
-        static_cast<Uint8>( std::min( color.g *percent / 100, 0xFF ) ),
-        static_cast<Uint8>( std::min( color.b *percent / 100, 0xFF ) ),
-        color.a
-    };
-}
-
-inline SDL_Color mix_colors( const SDL_Color &first, const SDL_Color &second, int second_percent )
-{
-    if( second_percent <= 0 ) {
-        return first;
-    }
-
-    if( second_percent >= 100 ) {
-        return second;
-    }
-
-    const int first_percent = 100 - second_percent;
-
-    return SDL_Color{
-        static_cast<Uint8>( std::min( ( first.r * first_percent + second.r * second_percent ) / 100, 0xFF ) ),
-        static_cast<Uint8>( std::min( ( first.g * first_percent + second.g * second_percent ) / 100, 0xFF ) ),
-        static_cast<Uint8>( std::min( ( first.b * first_percent + second.b * second_percent ) / 100, 0xFF ) ),
-        static_cast<Uint8>( std::min( ( first.a * first_percent + second.a * second_percent ) / 100, 0xFF ) )
-    };
-}
+SDL_Color mix_colors( const SDL_Color &first, const SDL_Color &second, int second_percent );
 
 inline bool is_black( const SDL_Color &color )
 {
@@ -68,91 +34,17 @@ inline Uint8 average_pixel_color( const SDL_Color &color )
     return 85 * ( color.r + color.g + color.b ) >> 8; // 85/256 ~ 1/3
 }
 
-inline SDL_Color color_pixel_grayscale( const SDL_Color &color )
-{
-    if( is_black( color ) ) {
-        return color;
-    }
+SDL_Color color_pixel_grayscale( const SDL_Color &color );
 
-    const Uint8 av = average_pixel_color( color );
-    const Uint8 result = std::max( av * 5 >> 3, 0x01 );
+SDL_Color color_pixel_nightvision( const SDL_Color &color );
 
-    return { result, result, result, color.a };
-}
+SDL_Color color_pixel_overexposed( const SDL_Color &color );
 
-inline SDL_Color color_pixel_nightvision( const SDL_Color &color )
-{
-    const Uint8 av = average_pixel_color( color );
-    const Uint8 result = std::min( ( av * ( ( av * 3 >> 2 ) + 64 ) >> 8 ) + 16, 0xFF );
+SDL_Color color_pixel_darken( const SDL_Color &color );
 
-    return {
-        static_cast<Uint8>( result >> 2 ),
-        static_cast<Uint8>( result ),
-        static_cast<Uint8>( result >> 3 ),
-        color.a
-    };
-}
+SDL_Color color_pixel_sepia( const SDL_Color &color );
 
-inline SDL_Color color_pixel_overexposed( const SDL_Color &color )
-{
-    const Uint8 av = average_pixel_color( color );
-    const Uint8 result = std::min( 64 + ( av * ( ( av >> 2 ) + 0xC0 ) >> 8 ), 0xFF );
-
-    return {
-        static_cast<Uint8>( result >> 2 ),
-        static_cast<Uint8>( result ),
-        static_cast<Uint8>( result >> 3 ),
-        color.a
-    };
-}
-
-inline SDL_Color color_pixel_darken( const SDL_Color &color )
-{
-    if( is_black( color ) ) {
-        return color;
-    }
-
-    // 85/256 ~ 1/3
-    return {
-        std::max<Uint8>( 85 * color.r >> 8, 0x01 ),
-        std::max<Uint8>( 85 * color.g >> 8, 0x01 ),
-        std::max<Uint8>( 85 * color.b >> 8, 0x01 ),
-        color.a
-    };
-
-}
-
-inline SDL_Color color_pixel_sepia( const SDL_Color &color )
-{
-    if( is_black( color ) ) {
-        return color;
-    }
-
-    /*
-     *  Objective is to provide a gradient between two color points
-     *  (sepia_dark and sepia_light) based on the grayscale value.
-     *  This presents an effect intended to mimic a faded sepia photograph.
-     */
-
-    const SDL_Color sepia_dark = { 39, 23, 19, color.a};
-    const SDL_Color sepia_light = { 241, 220, 163, color.a};
-
-    const Uint8 av = average_pixel_color( color );
-    const float gammav = 1.6;
-    const float pv = av / 255.0;
-    const Uint8 finalv =
-        std::min( static_cast<int>( std::round( std::pow( pv, gammav ) * 150 ) ), 100 );
-
-    return mix_colors( sepia_dark, sepia_light, finalv );
-}
-
-inline SDL_Color color_pixel_z_overlay( const SDL_Color &color )
-{
-    if( static_z_effect ) {
-        return mix_colors( color, {128, 255, 255, color.a}, color.a / 8 );
-    }
-    return { 128, 255, 255, color.a };
-}
+SDL_Color color_pixel_z_overlay( const SDL_Color &color );
 
 SDL_Color curses_color_to_SDL( const nc_color &color );
